Increase bundle version
The original MachOView (v2.4) by psaghelyi on SourceForge has the bundle ID `9201`, and that leads to problems that automated app updaters won't recognize e.g. your v2.6.1 as an update. So I increased the bundle ID from `271` to `10271`, keeping the original numbering scheme (271), but adding 10,000 to put it safely ahead of the original v2.4.